LayoutTests/http/tests/security/xssAuditor/regress-167121-expected.txt

CONSOLE MESSAGE: line 2: .../security/xssAuditor/regress-167121.html
This tests that the URL of an iframe whose page triggered a full page block is not "data:,".

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
